Backport: Modernize Platform.h (2) - noexcept, const, constexpr. ColourDesired is an int instead of long for consistency over different platforms. Changes made to Point, PRectangle, and ColourDesired. RoundXYPosition removed.
Backport of changeset 6939:7441dcb96e6a.

+++ cocoa/PlatCocoa.mm
@@ -276,7 +276,7 @@ void SurfaceImpl::PenColour(ColourDesired fore)
 {
   if (gc)
   {
-    ColourDesired colour(fore.AsLong());
+    ColourDesired colour(fore.AsInteger());
 
     // Set the Stroke color to match
     CGContextSetRGBStrokeColor(gc, colour.GetRed() / 255.0, colour.GetGreen() / 255.0,
@@ -290,7 +290,7 @@ void SurfaceImpl::FillColour(const ColourDesired& back)
 {
   if (gc)
   {
-    ColourDesired colour(back.AsLong());
+    ColourDesired colour(back.AsInteger());
 
     // Set the Fill color to match
     CGContextSetRGBFillColor(gc, colour.GetRed() / 255.0, colour.GetGreen() / 255.0,
@@ -911,7 +911,7 @@ void SurfaceImpl::DrawTextTransparent(PRectangle rc, Font &font_, XYPOSITION yba
                                       ColourDesired fore)
 {
 	CFStringEncoding encoding = EncodingFromCharacterSet(unicodeMode, FontCharacterSet(font_));
-	ColourDesired colour(fore.AsLong());
+	ColourDesired colour(fore.AsInteger());
 	CGColorRef color = CGColorCreateGenericRGB(colour.GetRed()/255.0,colour.GetGreen()/255.0,colour.GetBlue()/255.0,1.0);
 
 	QuartzTextStyle *style = TextStyleFromFont(font_);
